+++ src/Legacy/Legacy.Monolith/Catalog/Entities/Models/Brand.cs
@@ -1,0 +1,14 @@
+using Legacy.Monolith.Core.Entities.Models;
+
+namespace Legacy.Monolith.Catalog.Entities.Models;
+
+public class Brand : AuditableEntity
+{
+    public int Id { get; set; }
+
+    public string Name { get; set; } = default!;
+
+    public string PrimaryContactName { get; set; } = default!;
+
+    public string PrimaryContactEmail { get; set; } = default!;
+}
