+++ src/Legacy/Legacy.Monolith/Catalog/Entities/Models/Restriction.cs
@@ -1,0 +1,12 @@
+using Legacy.Monolith.Core.Entities.Models;
+
+namespace Legacy.Monolith.Catalog.Entities.Models;
+
+public class Restriction : AuditableEntity
+{
+    public int Id { get; set; }
+    public int ItemId { get; set; }
+    public Item Item { get; set; }
+    public bool IsRestricted { get; set; } = true;
+    public string Reason { get; set; }
+}
