+++ src/Legacy/Legacy.Monolith/Core/Entities/Models/AuditableEntity.cs
@@ -1,0 +1,7 @@
+namespace Legacy.Monolith.Core.Entities.Models;
+
+public class AuditableEntity
+{
+    public DateTime CreatedOn { get; set; }
+    public DateTime? ModifiedOn { get; set; }
+}
