--- src/Legacy/Legacy.Monolith/Entities/Configurations/BrandConfig.cs
@@ -1,25 +1,0 @@
-using Legacy.Monolith.Entities.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Legacy.Monolith.Entities.Configurations;
-
-public class BrandConfig : IEntityTypeConfiguration<Brand>
-{
-    public void Configure(EntityTypeBuilder<Brand> builder)
-    {
-        builder.HasKey(e => e.Id);
-
-        builder.Property(e => e.Name)
-            .IsRequired()
-            .HasMaxLength(64);
-
-        builder.Property(e => e.PrimaryContactName)
-            .IsRequired(false)
-            .HasMaxLength(100);
-
-        builder.Property(e => e.PrimaryContactEmail)
-            .IsRequired(false)
-            .HasMaxLength(100);
-    }
-}
