--- src/Legacy/Legacy.Monolith/Entities/Models/AuditableEntity.cs
@@ -1,7 +1,0 @@
-namespace Legacy.Monolith.Entities.Models;
-
-public class AuditableEntity
-{
-    public DateTime CreatedOn { get; set; }
-    public DateTime? ModifiedOn { get; set; }
-}
