--- src/Legacy/Legacy.Monolith/Entities/Models/Restriction.cs
@@ -1,10 +1,0 @@
-namespace Legacy.Monolith.Entities.Models;
-
-public class Restriction : AuditableEntity
-{
-    public int Id { get; set; }
-    public int ItemId { get; set; }
-    public Item Item { get; set; }
-    public bool IsRestricted { get; set; } = true;
-    public string Reason { get; set; }
-}
