+++ src/Legacy/Legacy.Monolith/Legacy.Monolith.csproj
@@ -30,4 +30,5 @@
     </ItemGroup>
 
 
+
 </Project>
